New changes

+++ Rogue.Server/Program.cs
@@ -79,17 +79,20 @@ namespace Rogue.Server
         [StartupInvoke("Synchronize Map", StartupInvokePriority.Primitive)]
         public static void SynchronizeMaps()
         {
-            string path = @"C:\Users\Skinz\Desktop\Heros-Die-Last\Rogue\bin\DesktopGL\AnyCPU\Debug\Maps";
-
-            foreach (var file in Directory.GetFiles(Environment.CurrentDirectory + MapRecord.MAPS_DIRECTORY))
+            if (Debugger.IsAttached)
             {
-                File.Delete(file);
-            }
+                string path = @"C:\Users\Skinz\Desktop\Heros-Die-Last\Rogue\bin\DesktopGL\AnyCPU\Debug\Maps";
 
-            foreach (var file in Directory.GetFiles(path))
-            {
-                var dest = Environment.CurrentDirectory + MapRecord.MAPS_DIRECTORY + Path.GetFileName(file);
-                File.Copy(file, dest);
+                foreach (var file in Directory.GetFiles(Environment.CurrentDirectory + MapRecord.MAPS_DIRECTORY))
+                {
+                    File.Delete(file);
+                }
+
+                foreach (var file in Directory.GetFiles(path))
+                {
+                    var dest = Environment.CurrentDirectory + MapRecord.MAPS_DIRECTORY + Path.GetFileName(file);
+                    File.Copy(file, dest);
+                }
             }
         }
